Fix my_basestat for high stats.
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@15803 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/textui/RuntimeLibrary.java
@@ -4321,15 +4321,15 @@ public abstract class RuntimeLibrary
 
 		if ( stat.equalsIgnoreCase( Stat.SUBMUSCLE.toString() ) )
 		{
-			return new Value( (int) KoLCharacter.getTotalMuscle() );
+			return new Value( KoLCharacter.getTotalMuscle() );
 		}
 		if ( stat.equalsIgnoreCase( Stat.SUBMYST.toString() ) )
 		{
-			return new Value( (int) KoLCharacter.getTotalMysticality() );
+			return new Value( KoLCharacter.getTotalMysticality() );
 		}
 		if ( stat.equalsIgnoreCase( Stat.SUBMOXIE.toString() ) )
 		{
-			return new Value( (int) KoLCharacter.getTotalMoxie() );
+			return new Value( KoLCharacter.getTotalMoxie() );
 		}
 
 		return DataTypes.ZERO_VALUE;
